Clean up tests some more

+++ src/Universalis.Application.Tests/Controllers/V1/Extra/Stats/MostRecentlyUpdatedItemsControllerTests.cs
@@ -25,7 +25,7 @@ public class MostRecentlyUpdatedItemsControllerTests
             await dbAccess.Push(74, new WorldItemUpload
             {
                 WorldId = 74,
-                ItemId = (int)itemId,
+                ItemId = itemId,
                 LastUploadTimeUnixMilliseconds = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds(),
             });
         }
